+++ CHANGELOG.md
@@ -47,6 +47,7 @@ UIIN-3437.
 * Add missing permissions for instance view. Refs UIIN-3472.
 * Fix quick instance export. Refs UIIN-3504.
 * Handle saving items order on dragEnd event. Refs UIIN-3454.
+* Show order cells as text inputs when invoking “Move items within an instance” action. Refs UIIN-3486.
 
 ## [13.0.10](https://github.com/folio-org/ui-inventory/tree/v13.0.10) (2025-09-01)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.9...v13.0.10)